fix: remove need for react import in component
React import was needed because jsx runtime is not automatically
injected by tsup. This could've been solved by setting `jsx:
'react-jsx'` in `tsconfig` but that adds `react/jsx-runtime` as a
dependency in the bundle. Instead, we set the esbuild option `jsx =
'automatic'` which handles this by automatically adding react import
statements where needed